Add generatePaymentToken option to checkout
This commit introduces a new option when creating a payment order, called "generatePaymentToken". Users will now see this option in the checkout area, providing the ability to generate a payment token for their transactions. The associated logic has been implemented in the CheckOut Controller and reflected in the relevant methods.